Move presentation tests to their own module
